chore: move entry package inside client
