+++ src/main/java/io/odpf/depot/redis/client/entry/RedisListEntry.java
@@ -1,4 +1,4 @@
-package io.odpf.depot.redis.entry;
+package io.odpf.depot.redis.client.entry;
 
 import io.odpf.depot.metrics.Instrumentation;
 import io.odpf.depot.redis.client.response.RedisClusterResponse;
@@ -16,27 +16,25 @@ import redis.clients.jedis.exceptions.JedisException;
  */
 @AllArgsConstructor
 @EqualsAndHashCode
-public class RedisHashSetFieldEntry implements RedisEntry {
-
+public class RedisListEntry implements RedisEntry {
     private final String key;
-    private final String field;
     private final String value;
     @EqualsAndHashCode.Exclude
     private final Instrumentation instrumentation;
 
     @Override
     public RedisStandaloneResponse send(Pipeline jedisPipelined, RedisTtl redisTTL) {
-        instrumentation.logDebug("key: {}, field: {}, value: {}", key, field, value);
-        Response<Long> response = jedisPipelined.hset(key, field, value);
+        instrumentation.logDebug("key: {}, value: {}", key, value);
+        Response<Long> response = jedisPipelined.lpush(key, value);
         redisTTL.setTtl(jedisPipelined, key);
         return new RedisStandaloneResponse(response);
     }
 
     @Override
     public RedisClusterResponse send(JedisCluster jedisCluster, RedisTtl redisTTL) {
-        instrumentation.logDebug("key: {}, field: {}, value: {}", key, field, value);
+        instrumentation.logDebug("key: {}, value: {}", key, value);
         try {
-            Long response = jedisCluster.hset(key, field, value);
+            Long response = jedisCluster.lpush(key, value);
             redisTTL.setTtl(jedisCluster, key);
             return new RedisClusterResponse(response.toString(), false);
         } catch (JedisException e) {
@@ -46,6 +44,6 @@ public class RedisHashSetFieldEntry implements RedisEntry {
 
     @Override
     public String toString() {
-        return String.format("RedisHashSetFieldEntry Key %s, Field %s, Value %s", key, field, value);
+        return String.format("RedisListEntry: Key %s, Value %s", key, value);
     }
 }
