+++ src/main/java/io/odpf/depot/redis/parsers/RedisEntryParser.java
@@ -2,7 +2,7 @@ package io.odpf.depot.redis.parsers;
 
 import io.odpf.depot.message.OdpfMessageSchema;
 import io.odpf.depot.message.ParsedOdpfMessage;
-import io.odpf.depot.redis.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
 
 import java.util.List;
