+++ src/main/java/io/odpf/depot/redis/parsers/RedisHashSetEntryParser.java
@@ -4,8 +4,8 @@ import io.odpf.depot.message.OdpfMessageSchema;
 import io.odpf.depot.message.ParsedOdpfMessage;
 import io.odpf.depot.metrics.Instrumentation;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.redis.entry.RedisEntry;
-import io.odpf.depot.redis.entry.RedisHashSetFieldEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisHashSetFieldEntry;
 
 import java.util.List;
 import java.util.Map;
