+++ src/main/java/io/odpf/depot/redis/parsers/RedisListEntryParser.java
@@ -1,13 +1,12 @@
 package io.odpf.depot.redis.parsers;
 
 
-import io.odpf.depot.config.RedisSinkConfig;
 import io.odpf.depot.message.OdpfMessageSchema;
 import io.odpf.depot.message.ParsedOdpfMessage;
 import io.odpf.depot.metrics.Instrumentation;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.redis.entry.RedisEntry;
-import io.odpf.depot.redis.entry.RedisListEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisListEntry;
 
 import java.util.Collections;
 import java.util.List;
