+++ src/main/java/io/odpf/depot/redis/parsers/RedisParser.java
@@ -6,7 +6,7 @@ import io.odpf.depot.error.ErrorType;
 import io.odpf.depot.exception.ConfigurationException;
 import io.odpf.depot.exception.DeserializerException;
 import io.odpf.depot.message.*;
-import io.odpf.depot.redis.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
 import io.odpf.depot.redis.record.RedisRecord;
 import lombok.AllArgsConstructor;
