+++ src/test/java/io/odpf/depot/redis/parsers/RedisHashSetEntryParserTest.java
@@ -10,8 +10,8 @@ import io.odpf.depot.message.ParsedOdpfMessage;
 import io.odpf.depot.message.SinkConnectorSchemaMessageMode;
 import io.odpf.depot.message.proto.ProtoOdpfMessageParser;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.redis.entry.RedisEntry;
-import io.odpf.depot.redis.entry.RedisHashSetFieldEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisHashSetFieldEntry;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
