+++ src/test/java/io/odpf/depot/redis/parsers/RedisKeyValueEntryParserTest.java
@@ -7,8 +7,8 @@ import io.odpf.depot.exception.ConfigurationException;
 import io.odpf.depot.message.*;
 import io.odpf.depot.message.proto.ProtoOdpfMessageParser;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.redis.entry.RedisEntry;
-import io.odpf.depot.redis.entry.RedisKeyValueEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisKeyValueEntry;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
