+++ src/test/java/io/odpf/depot/redis/parsers/RedisListEntryParserTest.java
@@ -7,8 +7,8 @@ import io.odpf.depot.exception.ConfigurationException;
 import io.odpf.depot.message.*;
 import io.odpf.depot.message.proto.ProtoOdpfMessageParser;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.redis.entry.RedisEntry;
-import io.odpf.depot.redis.entry.RedisListEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisListEntry;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
