+++ src/test/java/io/odpf/depot/redis/record/RedisRecordTest.java
@@ -4,7 +4,7 @@ import io.odpf.depot.error.ErrorInfo;
 import io.odpf.depot.error.ErrorType;
 import io.odpf.depot.redis.client.response.RedisClusterResponse;
 import io.odpf.depot.redis.client.response.RedisStandaloneResponse;
-import io.odpf.depot.redis.entry.RedisEntry;
+import io.odpf.depot.redis.client.entry.RedisEntry;
 import io.odpf.depot.redis.ttl.RedisTtl;
 import org.junit.Assert;
 import org.junit.Test;
